Split custom losses into seperate manageable files
