+++ reports/typesetting/sections/related_work.tex
@@ -22,7 +22,7 @@
 
 \IEEEPARstart{}{} Physics informed learning methods have emerged as a promising approach in a wide variety of applications by integrating the underlying physical principles into the learning process. The broader research community has been lauding over the introduction of physics informed machine learning (PIML) methodologies since they were introduced by Raissi \cite{raissi2017physics}. Within their seminal paper, hidden fluid flows were learned by a PINN showcasing their power to generalise well in the presence of noisy and low fidelity information \cite{raissi2018hidden}. This spawned multiple branches for PIML to explore, most of which have found PIML methods have increased the powers of deep learning methods, especially in a small data paradigm \cite{karniadakis2021physics}. 
 
-\IEEEPARstart{}{} Recently the term physics-informed computer vision (PICV) has been coined to describe the emerging domain of applying physics-informed models with computer vision tasks \cite{banerjee2023physics}. Physics-informed computer vision models exploit the benefits of having information gained from the underlying physics of the system and data to validate it's inference. The PICV domain is still in it's infancy and has many opportunities for rapid advancement. Examples of recent PICV applications include; superesolution \cite{arora2022spatio}, image generation \cite{zheng2020physics}, image reconstruction \cite{chu2022physics}, image segmentation \cite{jenkins2020physics, cciccek20163d} and crowd analysis \cite{behera2021pidlnet}. One such notable application includes inferring fluid flows from 4D-MRI imaging for the purposes of understanding arterial wall stress leveraging the well known Navier Stokes equations for incompressible fluid flows \cite{raissi2020hidden}. Estimations of the fluid flow vector field allows for the pressure and velocities to be estimated for advanced non-invasive diagnositcs for vulnerable patients. There is a strong link between \cite{raissi2020hidden} and learning of optical flow, however, it must be noted that as of yet we have not found any literature that directly learns an optical flow estimation using a PIML method. The links between physics-informed learning and algorithm unrolling are begining to strengthen, and our work aids to justify the link between the two. With optimisation proceedures directly induced into the deep learning network, the argument is that the architecture can resolve the physics of the system and make effective use of this information to perform it's designed task.
+\IEEEPARstart{}{} Recently the term physics-informed computer vision (PICV) has been coined to describe the emerging domain of applying physics-informed models with computer vision tasks \cite{banerjee2023physics}. Physics-informed computer vision models exploit the benefits of having information gained from the underlying physics of the system and data to validate it's inference. The PICV domain is still in it's infancy and has many opportunities for rapid advancement. Examples of recent PICV applications include; superesolution \cite{arora2022spatio}, image generation \cite{zheng2020physics}, image reconstruction \cite{chu2022physics}, image segmentation \cite{jenkins2020physics, cciccek20163d} and crowd analysis \cite{behera2021pidlnet}. One such notable application includes inferring fluid flows from 4D-MRI imaging for the purposes of understanding arterial wall stress leveraging the well known Navier Stokes equations for incompressible fluid flows \cite{raissi2020hidden}. Estimations of the fluid flow vector field allows for the pressure and velocities to be estimated for advanced non-invasive diagnositcs for vulnerable patients. There is a strong link between \cite{raissi2020hidden} and learning of optical flow, however, it must be noted that as of yet we have not found any literature that directly learns an optical flow estimation using a PIML method. The links between physics-informed learning and algorithm unrolling are beginning to strengthen, and our work aids to justify the link between the two. With optimisation proceedures directly induced into the deep learning network, the argument is that the architecture can resolve the physics of the system and make effective use of this information to perform it's designed task.
 
 \IEEEPARstart{}{} Algorithm unrolling has emerged as part of the contemporary computer vision research community and has impacted the intersection of iterative optimization techniques and deep learning, increasing both efficiency and model performance. The concept, pioneered by Gregor and LeCun with the introduction of the Learned Iterative Shrinkage-Thresholding Algorithm (LISTA) \cite{gregor2010learning}, where this approach has been effectively employed in various domains, such as compressive sensing MRI \cite{yang2016deep}, and continues to evolve. Widespread adoption is evidenced in a review by Monga et al. \cite{monga2021algorithm}. The review demonstrated the effectiveness that unrolled optimisation algorithms have in solving inverse problems and hold substantial benefits for model explainability. Additional benefits are realised by inducing prior knowledge into deep learning architectures, fullfilling the inductive bias definition of physics-informed learning, allowing for a principled approach to modelling the problem space with deep learning architectures that have prior understanding. Our work builds on these advancements by proposing a novel physics-informed algorithm unrolled deep learning architecture, specifically tailored for optical flow estimation. By unrolling gradient descent and proximal gradient descent, our method embeds knowledge of the physical constraints directly into the learning process, aiming to achieve enhanced generalization and accuracy even in data-scarce scenarios.
